requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
config==0.5.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.4
django-filter==22.1
django-jazzmin==2.6.0
django-phonenumber-field==7.0.1
django-phonenumbers==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==22.0
phonenumbers==8.13.2
Pillow==9.3.0
PyJWT==2.6.0
python-decouple==3.6
pytz==2022.7
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13

